finished - search indeed and linkedin jobs
