+++ linkedin.py
@@ -1,21 +1,18 @@
 #!/usr/bin/env python
 import json
 import logging
-import os
 import time
 from collections import OrderedDict, deque
-from datetime import datetime
-from os import path, makedirs
 from random import randint
 
-import pandas as pd
 from bs4 import BeautifulSoup
-from selenium import webdriver
 from selenium.webdriver.common.by import By
 from selenium.webdriver.common.keys import Keys
 
+from base_search import SearchJobs
 
-class SearchLinkedin:
+
+class SearchLinkedin(SearchJobs):
 
     def __init__(self, keywords, location, time_str, driver_path=None):
         """
@@ -34,34 +31,6 @@ class SearchLinkedin:
         self.driver_path = driver_path
         self.driver = None
 
-    @property
-    def out_file(self):
-        """ Output data file """
-        if len(self.location.split(' ')) == 1:
-            location = self.location.lower()
-        else:
-            location = ''.join([word[0].lower() for word in self.location.split(' ')])
-        return path.join('data', self.time_str + f'_{location}_{self.job_type}.csv')
-
-    @property
-    def job_type(self):
-        """ first letter of each keywords"""
-        return ''.join([word[0].lower() for word in self.keywords.split(' ')])
-
-    def init_webdriver(self):
-        """
-        Initialize a Chrome webdriver.
-        """
-        logging.info(f"Initialize a Chrome driver.")
-        if self.driver_path:
-            self.driver = webdriver.Chrome(self.driver_path)
-        else:
-            self.driver = webdriver.Chrome()
-        self.driver.implicitly_wait(10)  # seconds
-        time.sleep(0.5)
-        self.driver.maximize_window()
-        time.sleep(0.5)
-
     def get_credentials(self):
         """ Load login account and password from a local file """
         logging.info(f"Load a configuration file for job search.")
@@ -70,7 +39,7 @@ class SearchLinkedin:
         self.email = data['email']
         self.password = data['password']
 
-    def login_linkedin(self):
+    def login(self):
         """This function logs into your personal LinkedIn profile"""
         self.get_credentials()
 
@@ -176,7 +145,7 @@ class SearchLinkedin:
             return job_element.get_attribute('href').split('?')[0]
 
         result_jobs = OrderedDict()  # all jobs: {link: element}
-        link_elements = self.driver.find_elements(By.XPATH, "//a")
+        link_elements = self.driver.find_elements(By.XPATH, "//div/a")
         q_elements = deque([a for a in link_elements if is_job_link(a)])
         q_links = deque([get_href_link(a) for a in q_elements])
         while len(q_elements):
@@ -191,7 +160,7 @@ class SearchLinkedin:
 
                 # scroll to this job element and find all link elements
                 self.driver.execute_script("arguments[0].scrollIntoView();", job_element)
-                time.sleep(1)
+                time.sleep(randint(1, 3) * 0.1)
                 new_link_elements = self.driver.find_elements(By.XPATH, "//a")
 
                 # add new jobs which are not in the queue or in the result to the queue
@@ -217,7 +186,6 @@ class SearchLinkedin:
             job_element.click()
             time.sleep(randint(1, 3))
             soup = BeautifulSoup(self.driver.page_source, 'html.parser')
-            time.sleep(randint(1, 2) * 0.5)
 
             try:
                 job_title = soup.find('span', {'class': 'job-details-jobs-unified-top-card__job-title-link'}).get_text(
@@ -280,246 +248,3 @@ class SearchLinkedin:
             if page_num > 40:
                 logging.info(f"Exit as finished {page_num} job pages.")
                 break
-
-    def save_results(self, data):
-        """
-        Find interested jobs from job data and save it to a local csv file
-        :return:
-        """
-        df = pd.DataFrame(data)
-        if not path.exists(self.out_file):
-            logging.info(f"Save all jobs on this page to {self.out_file}")
-            df.to_csv(self.out_file, index=True)
-        else:
-            logging.info(f"Append all jobs on this page to {self.out_file}")
-            df.to_csv(self.out_file, mode='a', index=True, header=False)
-
-    def close_session(self):
-        """This function closes the actual session"""
-        logging.info('Close this session!')
-        self.driver.close()
-        time.sleep(1)
-
-    def run(self):
-        """ Search jobs and save results """
-        logging.info("Start...")
-        self.init_webdriver()
-        self.login_linkedin()
-        self.search_jobs()
-        self.filter()
-        self.scrape_jobs()
-        self.close_session()
-        logging.info("All done!")
-
-
-def config_log(log_file, level=logging.INFO) -> None:
-    """
-    Configure a log file.
-    @param log_file: path to a log file
-    @param level: choices can be logging.DEBUG, logging.INFO, ...
-    """
-    log_dir = path.dirname(log_file)
-    if not path.exists(log_dir):
-        makedirs(log_dir)
-
-    # Remove all handlers from the root logger
-    root_logger = logging.getLogger()
-    for handler in root_logger.handlers:
-        logging.root.removeHandler(handler)
-
-    # Create a new log file
-    str_format = '%(asctime)s - %(levelname)s - %(message)s'
-    logging.basicConfig(filename=log_file, level=level, format=str_format)
-
-    # Define a Handler which writes messages or higher to the sys.stderr
-    console = logging.StreamHandler()
-    console.setLevel(logging.INFO)
-    console.setFormatter(logging.Formatter(str_format, "%Y-%m-%d %H:%M:%S"))
-    # add the handler to the root logger
-    logging.getLogger('').addHandler(console)
-
-
-def is_intern_job(title):
-    """ Is a job an intern job? """
-    return True if 'intern' in title else False
-
-
-def require_us_citizenship(description):
-    """ Is US Citizenship required """
-    # description is all lowercase.
-    citizen_keywords = ['u.s. citizen', 'us citizen', 'united states citizen']
-    for keywords in citizen_keywords:
-        if keywords in description:
-            return True
-    return False
-
-
-def is_geo_job(description):
-    """
-    Is it a Geo job?
-    :param description: lowercase string description
-    :return:
-    """
-    # a geo job should contain any of the following keywords
-    geo_keywords = ['remote sensing', 'earth observation']
-    for keywords in geo_keywords:
-        if keywords in description:
-            return True
-    # a job description contains 2+ of the following keywords, it's a geo job
-    cnt = 0
-    for keyword in ['satellite', 'geospatial', 'image', 'climate change']:
-        if keyword in description:
-            cnt += 1
-    return cnt >= 2
-
-
-def is_ai_job(description):
-    """
-    Is it an AI job?
-    :param description: lowercase string description
-    :return:
-    """
-    ai_keywords = ['deep learning', 'pytorch', 'tensorflow']
-    for keywords in ai_keywords:
-        if keywords in description:
-            return True
-    return False
-
-
-def is_cv_job(description):
-    """
-    Is it a computer vision job?
-    :param description: lowercase string description
-    :return:
-    """
-    if 'computer vision' in description:
-        return True
-    return False
-
-
-def is_rs_job(description):
-    """
-    Is it a recommender system job?
-    :param description: lowercase string description
-    :return:
-    """
-    rs_keywords = ['recommender system', 'recommendation system']
-    for keywords in rs_keywords:
-        if keywords in description:
-            return True
-    return False
-
-
-def is_auto_job(description):
-    """
-    Is it an autonomous driving job
-    :param description:
-    :return:
-    """
-    cnt = 0
-    for keyword in ['autonomous driving', 'self-driving', 'perception', 'lidar', 'radar', 'calibration', 'mapping']:
-        if keyword in description:
-            cnt += 1
-    # if the description contains 2+ of the keywords, it's considered as an autonomous driving job
-    return cnt >= 2
-
-
-def select_jobs(file_list):
-    """
-    Select geo_ai and cv_ai jobs from the file list
-    :param file_list: a list of output files.
-    :return:
-    """
-    # There are usually overlaps between SDS and MLE jobs. Merge the two and remove redundancies.
-    geoai_links = set()
-    cvai_links = set()
-    rsai_links = set()
-    auto_links = set()
-
-    def select_geo_cv_jobs(in_file):
-        """ Select geo_ai and cv_ai jobs """
-        df = pd.read_csv(in_file)
-        logging.info(f"{df.shape[0]} job entries in {in_file}.")
-        geoai_inds = []
-        cvai_inds = []
-        rsai_inds = []
-        auto_inds = []
-        for i in range(df.shape[0]):
-            row = df.iloc[i]
-            link, title, company, location = row['Link'], row['Title'], row['Company'], row['Location']
-            logging.info(f"{i + 1}/{df.shape[0]}, {link}, {title}, {company}, {location}")
-            description = row['Description'].lower()
-            if is_intern_job(row['Title'].lower()) or require_us_citizenship(description):
-                continue
-            is_ai = is_ai_job(description)
-            is_geo = is_geo_job(description)
-            is_cv = is_cv_job(description)
-            is_rs = is_rs_job(description)
-            is_auto = is_auto_job(description)
-
-            if is_ai and is_geo and (link not in geoai_links):
-                geoai_inds.append(i)
-                geoai_links.add(link)
-            if is_ai and is_cv and (link not in cvai_links):
-                cvai_inds.append(i)
-                cvai_links.add(link)
-            if is_ai and is_rs and (link not in rsai_links):
-                rsai_inds.append(i)
-                rsai_links.add(link)
-            if is_ai and is_auto and (link not in auto_links):
-                auto_inds.append(i)
-                auto_links.add(link)
-        return df, geoai_inds, cvai_inds, rsai_inds, auto_inds
-
-    geoai_dfs, cvai_dfs, rsai_dfs, auto_dfs = [], [], [], []
-    for in_file in file_list:
-        df, geoai_inds, cvai_inds, rsai_inds, auto_inds = select_geo_cv_jobs(in_file)
-        geoai_dfs.append(df.loc[geoai_inds])
-        cvai_dfs.append(df.loc[cvai_inds])
-        rsai_dfs.append(df.loc[rsai_inds])
-        auto_dfs.append(df.loc[auto_inds])
-
-    geoai_df = pd.concat(geoai_dfs)
-    cvai_df = pd.concat(cvai_dfs)
-    rsai_df = pd.concat(rsai_dfs)
-    auto_df = pd.concat(auto_dfs)
-    logging.info(f"The number GeoAI job entries: {geoai_df.shape[0]}.")
-    logging.info(f"The number CV_AI job entries: {cvai_df.shape[0]}.")
-    logging.info(f"The number RS_AI job entries: {rsai_df.shape[0]}.")
-    logging.info(f"The number AutoAI job entries: {auto_df.shape[0]}.")
-
-    base_name = '_'.join(path.basename(file_list[-1]).split("_")[0:2])
-    if geoai_df.shape[0] > 0:
-        geoai_file = path.join(path.dirname(file_list[-1]), base_name + '_geoai.csv')
-        geoai_df.to_csv(geoai_file, index=True)
-    if cvai_df.shape[0] > 0:
-        cvai_file = path.join(path.dirname(file_list[-1]), base_name + '_cvai.csv')
-        cvai_df.to_csv(cvai_file, index=True)
-    if rsai_df.shape[0] > 0:
-        rsai_file = path.join(path.dirname(file_list[-1]), base_name + '_rsai.csv')
-        rsai_df.to_csv(rsai_file, index=True)
-    if auto_df.shape[0] > 0:
-        auto_file = path.join(path.dirname(file_list[-1]), base_name + '_autoai.csv')
-        auto_df.to_csv(auto_file, index=True)
-
-    # delete the files
-    for in_file in file_list:
-        os.remove(in_file)
-    logging.info("Done!")
-
-
-if __name__ == '__main__':
-    # configure a log file
-    time_str = datetime.now().strftime("%Y%m%dH%H")
-    config_log(path.join('data/logs', time_str + '.log'))
-
-    job_titles = ["Machine Learning Engineer", "Senior Data Scientist", "Research Scientist"]
-    locations = ["United States", "Canada"]
-    for location in locations:
-        out_files = []
-        for job_title in job_titles:
-            bot = SearchLinkedin(job_title, location, time_str)
-            bot.run()
-            out_files.append(bot.out_file)
-        logging.info("Select interesting jobs form the search list.")
-        select_jobs(out_files)
